spark-application/requirements.txt
pyspark==3.2.0
delta-spark==1.2.1
pytest==7.4.4
chispa==0.10.0